pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lmx"
version = "0.2.0"
description = "The smart model picker for LLMs — tell me which one to use"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["llm", "ai", "cli", "cost-optimization", "model-selection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.0.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiosqlite>=0.20.0",
    "platformdirs>=4.0.0",
    "python-dateutil>=2.8.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.9.0",
]

[project.scripts]
lmx = "lmx.cli:app"

[project.urls]
Homepage = "https://github.com/my-ai-stack/lmx"
Repository = "https://github.com/my-ai-stack/lmx"

[tool.hatch.build.targets.wheel]
packages = ["src/lmx"]

[tool.ruff]
line-length = 100
target-version = "py310"
